Rename and refactor Modal to allow consumers to update without breaking.
