+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-<a name="4.0.0"></a>
-## [4.0.0](https://github.com/gabyvs/ng2-ue-utils/compare/3.2.0...4.0.0) (2017-02-21)
+<a name="3.3.0"></a>
+## [4.0.0](https://github.com/gabyvs/ng2-ue-utils/compare/3.2.0...3.3.0) (2017-02-21)
 
-### Breaking Changes
+### Features
 
-* **Modal:** Made modal more flexible by allowing consumers to define contents via transcludes and emitting whenever user takes action to close modal, regardless of the result of that action.
-* **BinaryChoiceModal:** Implementation of Modal which emits a boolean to represent user intent: `true` for submission, `false` for cancellation.
+* **BaseModal:** A more flexible modal which allows consumers to define contents via transcludes and emits when the modal closes.
+
+### Changes
+
+* **Modal:** Refactored Modal to use the new BaseModal.  Added optional input `cancelLabel` for defining the text content of the cancel button.
 
 <a name="3.2.0"></a>
 ## [3.2.0](https://github.com/gabyvs/ng2-ue-utils/compare/3.1.2...3.2.0) (2017-02-17)
