+++ demo/demo.html
@@ -200,42 +200,42 @@
     <h2>Modal: Component</h2>
     <div class="explanation">A modal with an html template and action bindings which you can define in your application.  For convenience, we offer a base <strong>Modal</strong> which simply emits an undefined event when it closes, as well as a <strong>BinaryChoiceModal</strong> which emits with the result of the user's action.</div>     
     
-        <modal #myModal [modalClass]="classForE2E" (emitClose)="baseModalClose()">
+        <modal-base #myModal [modalClass]="classForE2E" (emitClose)="baseModalClose()">
             <span modal-title>Do a thing?</span>
             <p modal-body>Are you sure you want to do this thing?  This thing cannot be undone.  Oh, the shame...</p>
             <div modal-footer>
                 <button>Not at all, dude</button>
                 <button (click)="baseModalSubmit(myModal)">Yeah, totally!</button>
             </div>
+        </modal-base>
+        <modal #myBinaryChoiceModal
+                             title="Do something?"
+                             cancelLabel="Don't do it"
+                             submitLabel="Do it"
+                             (emitSubmit)="binaryChoiceModalHandler($event)">
+            <p>Should I do something?</p>
         </modal>
-        <binary-choice-modal #myBinaryChoiceModal
-                             titleText="Do something?"
-                             bodyText="Do it or don't do it?"
-                             cancelText="Don't do it"
-                             confirmText="Do it"
-                             (modalResult)="binaryChoiceModalHandler($event)">
-        </binary-choice-modal>
         
         <div class="elsewhereOnThePage">
-            <button (click)="openModal(myModal)">Open modal</button>
+            <button (click)="openModal(myModal)">Open Modal Base</button>
         </div>
         <div *ngIf="baseModalCloseCount" class="resultBlock">
-            <span class="codeBlock">Modal was <strong>closed</strong> {{ modalCountText('baseModalCloseCount') }}</span>
+            <span class="codeBlock">Modal was <strong class="closed">closed</strong> {{ modalCountText('baseModalCloseCount') }}</span>
         </div>
         <div *ngIf="baseModalCloseCount" class="resultBlock">
-            <span class="codeBlock">Modal action was <strong>submitted</strong> {{ modalCountText('baseModalSubmitCount') }}</span>
+            <span class="codeBlock">Modal action was <strong class="submitted">submitted</strong> {{ modalCountText('baseModalSubmitCount') }}</span>
         </div>
         <div class="elsewhereOnThePage">
-            <button (click)="openModal(myBinaryChoiceModal)">Open binary choice modal</button>
+            <button (click)="openModal(myBinaryChoiceModal)">Open Modal</button>
         </div>
         <div *ngIf="binaryModalCloseCount" class="resultBlock">
-            <span class="codeBlock">Modal was <strong>closed</strong> {{ modalCountText('binaryModalCloseCount') }}</span>
+            <span class="codeBlock">Modal was <strong class="closed">closed</strong> {{ modalCountText('binaryModalCloseCount') }}</span>
         </div>
         <div *ngIf="binaryModalCloseCount" class="resultBlock">
-            <span class="codeBlock">Modal action was <strong>submitted</strong> {{ modalCountText('binaryModalSubmitCount') }}</span>
+            <span class="codeBlock">Modal action was <strong class="submitted">submitted</strong> {{ modalCountText('binaryModalSubmitCount') }}</span>
         </div>
         <div *ngIf="binaryModalCloseCount" class="resultBlock">
-            <span class="codeBlock">Modal action was <strong>cancelled</strong> {{ modalCountText('binaryModalCancelCount') }}</span>
+            <span class="codeBlock">Modal action was <strong class="cancelled">cancelled</strong> {{ modalCountText('binaryModalCancelCount') }}</span>
         </div>
 </div>
