--- src/components/binary-choice-modal/binary-choice-modal.html
@@ -1,8 +1,0 @@
-<modal (emitClose)="emitResult()" [modalClass]="modalClass">
-    <span modal-title>{{ titleText }}</span>
-    <p modal-body>{{ bodyText }}</p>
-    <div modal-footer>
-        <button>{{ cancelText }}</button>
-        <button (click)="modalSubmit()">{{ confirmText }}</button>
-    </div>
-</modal>
