+++ src/components/modal-base/modal-base.less
@@ -1,0 +1,26 @@
+@import '../../common/variables';
+.modal {
+    color: @table-color;
+    font-weight: 300;
+    .modal-header {
+        border-top: 4px solid @red-color;
+        border-radius: 4px;
+        border-bottom: none;
+        padding: 15px 15px 10px 15px;
+        .modal-title {
+            font-size: 22px;
+            font-weight: lighter;
+        }
+        .close {
+            font-size: 32px;
+            font-weight: lighter;
+        }
+    }
+    .modal-body {
+        padding: 10px 15px;
+    }
+    .modal-footer {
+        border-top: none;
+        padding: 10px 15px 15px 15px;
+    }
+}
