+++ src/components/modal/modal.html
@@ -1,24 +1,10 @@
-<div class="modal"
-     tabindex="-1" role="dialog"
-     #modal
-     (keyup.esc)="close()"
-     (click)="close()"
-     [ngClass]="{ in: isOpened, fade: isOpened }"
-     [ngStyle]="{ display: isOpened ? 'block' : 'none' }">
-    <div [class]="combinedModalClass">
-        <div class="modal-content" *ngIf="isOpened" tabindex="0">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal"><span aria-hidden="true">&times;</span></button>
-                <h4 class="modal-title">
-                    <ng-content select="[modal-title]"></ng-content>
-                </h4>
-            </div>
-            <div class="modal-body">
-                <ng-content select="[modal-body]"></ng-content>
-            </div>
-            <div class="modal-footer">
-                <ng-content select="[modal-footer]"></ng-content>
-            </div>
-        </div>
+<modal-base (emitClose)="emit()" [modalClass]="modalClass">
+    <span modal-title>{{ title }}</span>
+    <span modal-body>
+        <ng-content></ng-content>
+    </span>
+    <div modal-footer>
+        <button class="btn btn-default">{{ cancelLabel }}</button>
+        <button class="btn btn-primary" (click)="modalSubmit()">{{ submitLabel }}</button>
     </div>
-</div>
+</modal-base>
