+++ src/ng2-ue-utils.module.ts
@@ -6,8 +6,8 @@ import { Filtering }                from './components/filtering/filtering';
 import { HintScroll }               from './components/hint-scroll/hint-scroll';
 import { ListHeaders }              from './components/list-headers/list-headers';
 import { LoadingDots }              from './components/loading-dots/loading-dots';
-import { BinaryChoiceModal }        from './components/binary-choice-modal/binary-choice-modal';
 import { Modal }                    from './components/modal/modal';
+import { ModalBase }                from './components/modal-base/modal-base';
 import { Notification }             from './components/notification/notification';
 import { Pagination }               from './components/pagination/pagination';
 import { Progress }                 from './components/progress/progress';
@@ -25,12 +25,12 @@ import { DropdownTriggerDirective } from './directives/dropdown/dropdown-trigger
 
 @NgModule({
     declarations: [
-        BinaryChoiceModal,
         Filtering,
         HintScroll,
         ListHeaders,
         LoadingDots,
         Modal,
+        ModalBase,
         Notification,
         Pagination,
         Progress,
@@ -52,7 +52,7 @@ import { DropdownTriggerDirective } from './directives/dropdown/dropdown-trigger
         ListHeaders,
         LoadingDots,
         Modal,
-        BinaryChoiceModal,
+        ModalBase,
         Notification,
         Pagination,
         Progress,
