+++ tsconfig.json
@@ -20,7 +20,7 @@
     "src/components/hint-scroll/hint-scroll.spec.ts",
     "src/components/list-headers/list-headers.spec.ts",
     "src/components/modal/modal.spec.ts",
-    "src/components/binary-choice-modal/binary-choice-modal.spec.ts",
+    "src/components/modal-base/modal-base.spec.ts",
     "src/components/notification/notification.spec.ts",
     "src/components/pagination/pagination.spec.ts",
     "src/components/progress/progress.spec.ts",
